Main/Medium+MediumFormat+GuestOSType+SystemPropertiesImpl+Console+Global: consistently use bytes as size units, forgotten const value in API, MaxVDISize method renamed to InfoVDSize, STDMETHOD macro usage fixes, whitespace cleanup
Frontends/VirtualBox+VBoxManage+VBoxShell: adapt to changed disk size units

Main/StorageControllerImpl: check the storage controller instance limit to avoid creating unusable VM configs, simplify unnecessarily complex code for querying the controller properties


git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@32531 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/include/Global.h
@@ -60,7 +60,7 @@ public:
         const uint32_t                 osHint;
         const uint32_t                 recommendedRAM;
         const uint32_t                 recommendedVRAM;
-        const uint32_t                 recommendedHDD;
+        const uint64_t                 recommendedHDD;
         const NetworkAdapterType_T     networkAdapterType;
         const uint32_t                 numSerialEnabled;
         const StorageControllerType_T  dvdStorageControllerType;
